update Github example
