--- Examples/GithubExample/Code/AppDelegate.swift
@@ -1,42 +1,0 @@
-//
-//  AppDelegate.swift
-//  GithubExample
-//
-//  Created by Oliver Krakora on 12.04.19.
-//  Copyright © 2019 aaa - all about apps GmbH. All rights reserved.
-//
-
-import UIKit
-import Fetch
-
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    var window: UIWindow?
-
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        let config = Config(baseURL: URL(string: "https://api.github.com")!,
-                            cache: MemoryCache(defaultExpiration: .seconds(3600)),
-                            cachePolicy: .cacheFirstNetworkAlways)
-        APIClient.shared.setup(with: config)
-        return true
-    }
-
-    func applicationWillResignActive(_ application: UIApplication) {
-
-    }
-
-    func applicationDidEnterBackground(_ application: UIApplication) {
-
-    }
-
-    func applicationWillEnterForeground(_ application: UIApplication) {
-    }
-
-    func applicationDidBecomeActive(_ application: UIApplication) {
-    }
-
-    func applicationWillTerminate(_ application: UIApplication) {
-    }
-
-}
