+++ Examples/GithubExample/GithubExample/Code/AppDelegate.swift
@@ -1,0 +1,38 @@
+//
+//  AppDelegate.swift
+//  GithubExample
+//
+//  Created by Stefan Wieland on 05.01.22.
+//
+
+import UIKit
+import Fetch
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        let config = Config(baseURL: URL(string: "https://api.github.com")!,
+                            cache: MemoryCache(defaultExpiration: .seconds(3600)),
+                            cachePolicy: .cacheFirstNetworkAlways)
+        APIClient.shared.setup(with: config)
+        return true
+    }
+
+    // MARK: UISceneSession Lifecycle
+
+    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
+        // Called when a new scene session is being created.
+        // Use this method to select a configuration to create the new scene with.
+        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
+    }
+
+    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
+        // Called when the user discards a scene session.
+        // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
+        // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
+    }
+
+
+}
+
