+++ Examples/GithubExample/GithubExample/Code/Models/Organization.swift
