+++ Examples/GithubExample/GithubExample/Code/Models/User.swift
