+++ Examples/GithubExample/GithubExample/Code/Utilities/Paging/PageProtocol.swift
