+++ Examples/GithubExample/GithubExample/Code/Utilities/Paging/PagedResource.swift
