+++ Examples/GithubExample/GithubExample/Code/Views/RepositoryCell.swift
