+++ Examples/GithubExample/GithubExample/Code/Views/UserCell.swift
